# Js list prepending optimisation