--- a/compiler-core/src/javascript/expression.rs
+++ b/compiler-core/src/javascript/expression.rs
@@ -141,16 +141,17 @@ impl<'module> Generator<'module> {
             TypedExpr::Int { value, .. } => Ok(int(value)),
             TypedExpr::Float { value, .. } => Ok(float(value)),
 
-            TypedExpr::List { elements, tail, .. } => {
-                self.tracker.list_used = true;
-                self.not_in_tail_position(|gen| {
-                    let tail = match tail {
-                        Some(tail) => Some(gen.wrap_expression(tail)?),
-                        None => None,
-                    };
-                    list(elements.iter().map(|e| gen.wrap_expression(e)), tail)
-                })
-            }
+            TypedExpr::List { elements, tail, .. } => self.not_in_tail_position(|gen| match tail {
+                Some(tail) => {
+                    gen.tracker.prepend_used = true;
+                    let tail = gen.wrap_expression(tail)?;
+                    prepend(elements.iter().map(|e| gen.wrap_expression(e)), tail)
+                }
+                None => {
+                    gen.tracker.list_used = true;
+                    list(elements.iter().map(|e| gen.wrap_expression(e)))
+                }
+            }),
 
             TypedExpr::Tuple { elems, .. } => self.tuple(elems),
             TypedExpr::TupleIndex { tuple, index, .. } => self.tuple_index(tuple, *index),
@@ -1179,7 +1180,6 @@ pub(crate) fn guard_constant_expression<'a>(
                 elements
                     .iter()
                     .map(|e| guard_constant_expression(assignments, tracker, e)),
-                None,
             )
         }
         Constant::Record { typ, name, .. } if typ.is_bool() && name == "True" => {
@@ -1234,10 +1234,7 @@ pub(crate) fn constant_expression<'a>(
 
         Constant::List { elements, .. } => {
             tracker.list_used = true;
-            list(
-                elements.iter().map(|e| constant_expression(tracker, e)),
-                None,
-            )
+            list(elements.iter().map(|e| constant_expression(tracker, e)))
         }
 
         Constant::Record { typ, name, .. } if typ.is_bool() && name == "True" => {
@@ -1354,20 +1351,22 @@ pub fn array<'a, Elements: IntoIterator<Item = Output<'a>>>(elements: Elements) 
     .group())
 }
 
-fn list<'a, I: IntoIterator<Item = Output<'a>>>(
-    elements: I,
-    tail: Option<Document<'a>>,
-) -> Output<'a>
+fn list<'a, I: IntoIterator<Item = Output<'a>>>(elements: I) -> Output<'a>
 where
     I::IntoIter: DoubleEndedIterator + ExactSizeIterator,
 {
     let array = array(elements);
-    if let Some(tail) = tail {
-        let args = [array, Ok(tail)];
-        Ok(docvec!["toList", call_arguments(args)?])
-    } else {
-        Ok(docvec!["toList(", array?, ")"])
-    }
+    Ok(docvec!["toList(", array?, ")"])
+}
+
+fn prepend<'a, I: IntoIterator<Item = Output<'a>>>(elements: I, tail: Document<'a>) -> Output<'a>
+where
+    I::IntoIter: DoubleEndedIterator + ExactSizeIterator,
+{
+    elements.into_iter().rev().fold(Ok(tail), |tail, element| {
+        let args = [element, tail];
+        Ok(docvec!["$prepend", call_arguments(args)?])
+    })
 }
 
 fn call_arguments<'a, Elements: IntoIterator<Item = Output<'a>>>(elements: Elements) -> Output<'a> {