… compiler-core/src/javascript/tests/snapshots/gleam_core__javascript__tests__lists__list_literals.snap