This PR changes the compiled JS to call a `prepend` function instead of using the `toList` function, to improve runtime performance. For example `[1, ..xs]` is now compiled to `$prepend(1, xs)` instead of `toList([1], xs)`, where `$prepend` is a thin wrapper over the list constructor.

Related issue #2562

# Benchmarks

A couple of benchmarks comparing the nightly build against this branch.

## gleam-stdlib "time gleam test"

Runtime of the tests of gleam-stdlib measured using `time gleam test` (3 runs).

### before

```
2.98s user 0.37s system 135% cpu 2.471 total
2.85s user 0.37s system 133% cpu 2.409 total
3.00s user 0.39s system 134% cpu 2.516 total
```

### after

```
2.37s user 0.35s system 132% cpu 2.048 total
2.33s user 0.37s system 134% cpu 2.009 total
2.38s user 0.36s system 134% cpu 2.034 total
```

## gleamy_bench example benchmark

Running the example benchmark from the gleamy_bench library.

### before

```
Input               Function           IPS         Min         P99
pre-sorted list     list.sort()     6.5211    146.2742    168.8623
reversed list       list.sort()     5.8963    160.0505    194.4547
```

### after

```
Input               Function           IPS         Min         P99
pre-sorted list     list.sort()     8.2688    113.6544    136.9221
reversed list       list.sort()     8.2111    112.7967    138.6488
```