@schurhammer 2024-02-29
Added to changelog

We've got this test already so I think its covered
```gleam
---
source: compiler-core/src/javascript/tests/lists.rs
expression: "\nfn go(x) {\n    []\n    [1]\n    [1, 2]\n    [1, 2, ..x]\n}\n"
---
import { toList, prepend as $prepend } from "../gleam.mjs";

function go(x) {
  toList([]);
  toList([1]);
  toList([1, 2]);
  return $prepend(1, $prepend(2, x));
}
```